requirements.txt
streamlit
chromadb
sentence-transformers
PyPDF2
requests
faiss-cpu
